refactor deps resolver
